resolve a warning about unsigned/signed comparison

+++ Source/automap.cpp
@@ -84,7 +84,7 @@ void InitAutomap()
 	pAFile = LoadFileInMem(mapData, &dwTiles);
 
 	dwTiles /= 2;
-	assert(dwTiles < lengthof(automaptype));
+	assert(dwTiles < (DWORD)lengthof(automaptype));
 	pTmp = pAFile;
 
 	for (i = 1; i <= dwTiles; i++) {
